+++ packages/browser/types/index.d.ts
@@ -1,10 +1,5 @@
 import {
   History,
-  Key,
-  Pathname,
-  Query,
-  Hash,
-  State,
   LocationDetails,
   PartialLocation,
   HickoryLocation,
@@ -16,14 +11,9 @@ export {
   HickoryLocation,
   PartialLocation,
   AnyLocation,
-  LocationDetails,
-  Key,
-  Pathname,
-  Query,
-  Hash,
-  State
+  LocationDetails
 };
 export interface Options extends RootOptions {
-  raw?: (pathname: Pathname) => Pathname;
+  raw?: (pathname: string) => string;
 }
 export default function Browser(options?: Options): History;