+++ packages/in-memory/types/index.d.ts
@@ -1,10 +1,5 @@
 import {
   History,
-  Key,
-  Pathname,
-  Query,
-  Hash,
-  State,
   LocationDetails,
   HickoryLocation,
   PartialLocation,
@@ -16,11 +11,6 @@ export {
   HickoryLocation,
   PartialLocation,
   AnyLocation,
-  Key,
-  Pathname,
-  Query,
-  Hash,
-  State,
   LocationDetails
 };
 export interface Options extends RootOptions {