+++ packages/root/src/types/location.ts
@@ -1,29 +1,23 @@
-export type Pathname = string;
-export type Query = any;
-export type Hash = string;
-export type State = any | null;
-export type Key = string;
-
 export interface LocationDetails {
-  query?: Query;
-  hash?: Hash;
-  state?: State;
+  query?: any;
+  hash?: string;
+  state?: any;
 }
 
 export interface PartialLocation extends LocationDetails {
-  pathname?: Pathname;
-  rawPathname?: Pathname;
+  pathname?: string;
+  rawPathname?: string;
 }
 
 // use HickoryLocation instead of Location to prevent
 // errors from colling with window.Location interface
 export interface HickoryLocation {
-  pathname: Pathname;
-  query: Query;
-  hash: Hash;
-  state?: State;
-  key: Key;
-  rawPathname: Pathname;
+  pathname: string;
+  query: any;
+  hash: string;
+  state?: any;
+  key: string;
+  rawPathname: string;
 }
 
 export type AnyLocation = HickoryLocation | PartialLocation;