+++ packages/root/types/types/keygen.d.ts
@@ -1,9 +1,8 @@
-import { Key } from "./location";
 export interface KeyMethods {
   keygen: KeyFns;
 }
 export interface KeyFns {
-  major(previous?: Key): Key;
-  minor(current: Key): Key;
-  diff(first: Key, second: Key): number;
+  major(previous?: string): string;
+  minor(current: string): string;
+  diff(first: string, second: string): number;
 }