+++ packages/root/types/types/location.d.ts
@@ -1,23 +1,18 @@
-export declare type Pathname = string;
-export declare type Query = any;
-export declare type Hash = string;
-export declare type State = any | null;
-export declare type Key = string;
 export interface LocationDetails {
-  query?: Query;
-  hash?: Hash;
-  state?: State;
+  query?: any;
+  hash?: string;
+  state?: any;
 }
 export interface PartialLocation extends LocationDetails {
-  pathname?: Pathname;
-  rawPathname?: Pathname;
+  pathname?: string;
+  rawPathname?: string;
 }
 export interface HickoryLocation {
-  pathname: Pathname;
-  query: Query;
-  hash: Hash;
-  state?: State;
-  key: Key;
-  rawPathname: Pathname;
+  pathname: string;
+  query: any;
+  hash: string;
+  state?: any;
+  key: string;
+  rawPathname: string;
 }
 export declare type AnyLocation = HickoryLocation | PartialLocation;