+++ packages/in-memory/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-* Update re-exported types.
+* Re-export `LocationDetails`.
 * Don't publish source maps.
 
 ## 1.0.0-beta.3